Split out gtk tests
