+++ tests/utils.py
@@ -21,13 +21,25 @@ def glib_policy():
     return GLibEventLoopPolicy()
 
 
+@pytest.fixture
+def gtk_policy():
+    from gbulb.gtk import GtkEventLoopPolicy
+    return GtkEventLoopPolicy()
+
+
 @pytest.yield_fixture(scope='function')
 def glib_loop():
     l = glib_policy().new_event_loop()
     setup_test_loop(l)
-
     yield l
-
     check_loop_failures(l)
+    l.close()
 
+
+@pytest.yield_fixture(scope='function')
+def gtk_loop():
+    l = gtk_policy().new_event_loop()
+    setup_test_loop(l)
+    yield l
+    check_loop_failures(l)
     l.close()
